Move pages to their own containers
