+++ sample/src/main/java/org/firezenk/kartographer/di/ApplicationComponent.kt
@@ -2,9 +2,9 @@ package org.firezenk.kartographer.di
 
 import dagger.Component
 import org.firezenk.kartographer.MainActivity
-import org.firezenk.kartographer.pages.Page1
+import org.firezenk.kartographer.pages.page1.Page1
 import org.firezenk.kartographer.pages.page2.Page2
-import org.firezenk.kartographer.pages.Page3
+import org.firezenk.kartographer.pages.page3.Page3
 import javax.inject.Singleton
 
 /**
